pyproject.toml
[project]
name = "fun-things"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "djangorestframework>=3.15.2",
    "requests>=2.32.3",
    "python-dotenv==1.0.1",
    "tqdm==4.67.1",
    "django-extensions==3.2.3",
    "django-cors-headers==4.6.0",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "distro==1.9.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "jiter==0.8.2",
    "openai==1.60.2",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "sniffio==1.3.1",
    "typing-extensions==4.12.2",
]

[build-system]
requires = [
    "setuptools",
    "wheel",
]
build-backend = "setuptools.build_meta"